Fix PriceRange-PopUp
