+++ src/App.css
@@ -36,3 +36,13 @@
 .host {
   color: grey;
 }
+
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.6); 
+  z-index: 1100;
+}
