+++ src/App.jsx
@@ -6,30 +6,17 @@ import CategoryTabs from "./components/CategoryTabs/CategoryTabs";
 import ProductCard from "./components/ProductCard/ProductCard";
 import CalendarToggle from "./components/calendarToggle/CalendarToggle";
 import { BASE_URL } from "./constants/constants";
+import PriceRangeModal from "./components/PriceRangeModal/PriceRangeModal";
 
 
 function App() {
   const [places, setPlaces] = useState([]);
   const [selectPlaceId, setSelectPlaceId] = useState(null);
   const [searchParams] = useSearchParams();
+  const [isModalOpen, setModalOpen] = useState(false);
+  const [histogramData, setHistogramData] = useState([]);
 
-  const histogramData = [
-    { from: 16, to: 23, count: 2 },
-    { from: 37, to: 44, count: 13 },
-    { from: 55, to: 63, count: 30 },
-    { from: 76, to: 84, count: 50 },
-    { from: 95, to: 103, count: 90 },
-    { from: 116, to: 123, count: 60 },
-    { from: 135, to: 143, count: 20 },
-    { from: 156, to: 156, count: 12 },
-    { from: 174, to: 182, count: 15 },
-    { from: 194, to: 202, count: 8 },
-    { from: 216, to: 224, count: 5 },
-    { from: 233, to: 242, count: 10 },
-    { from: 254, to: 261, count: 25 },
-    { from: 273, to: 281, count: 40 },
-    { from: 292, to: 304, count: 22 },
-  ];
+  const toggleModal = () => setModalOpen((prev) => !prev);
 
   useEffect(() => {
     axios
@@ -58,7 +45,7 @@ function App() {
         <CalendarToggle />
       </div>
       
-      <CategoryTabs />  
+      <CategoryTabs toggleModal={toggleModal} setHistogramData={setHistogramData} />
 
       <div className="grid">
         {places.map((place) => {
@@ -79,6 +66,14 @@ function App() {
           );
         })}
       </div>
+      {isModalOpen && (
+        <PriceRangeModal
+          isOpen={isModalOpen}
+          className="overlay"
+          onClose={toggleModal}
+          histogramData={histogramData}
+        />
+      )}
     </>
   );
 }
