+++ src/components/Header/Header.module.css
@@ -1,7 +1,7 @@
 .headerSectionContainer {
     width: 100%;
     position: relative;
-    z-index: 900;
+    z-index: 1000;
 }
 
 .header {
